Investment case study
Spark Funds, an asset management company, wants to make investments in a few companies
Spark Funds has two minor constraints for investments:
It wants to invest between 5 to 15 million USD per round of investment
It wants to invest only in English-speaking countries because of the ease of communication with the companies it would invest in.
We do an analysis on real investment data and our goal is to find the best Investment type, the countries most heavily invested in and the best sector to invest in.
The overall strategy is to invest where others are investing, implying that the 'best' sectors and countries are the ones 'where most investors are investing'.